+++ examples/simple-generator/CHANGELOG.md
@@ -1,5 +1,58 @@
 # @sap-ux/generator-simple-fe
 
+## 0.2.144
+
+### Patch Changes
+
+-   Updated dependencies [b63aea83]
+    -   @sap-ux/fiori-freestyle-writer@0.20.1
+    -   @sap-ux/fiori-elements-writer@0.13.1
+
+## 0.2.143
+
+### Patch Changes
+
+-   Updated dependencies [29e71f68]
+    -   @sap-ux/axios-extension@1.4.2
+    -   @sap-ux/system-access@0.2.1
+
+## 0.2.142
+
+### Patch Changes
+
+-   Updated dependencies [de0b13f2]
+    -   @sap-ux/fiori-elements-writer@0.13.0
+    -   @sap-ux/fiori-freestyle-writer@0.20.0
+
+## 0.2.141
+
+### Patch Changes
+
+-   Updated dependencies [44df3d5c]
+    -   @sap-ux/fiori-freestyle-writer@0.19.17
+    -   @sap-ux/fiori-elements-writer@0.12.27
+
+## 0.2.140
+
+### Patch Changes
+
+-   Updated dependencies [b69c21f3]
+    -   @sap-ux/fiori-elements-writer@0.12.26
+
+## 0.2.139
+
+### Patch Changes
+
+-   @sap-ux/fiori-elements-writer@0.12.25
+
+## 0.2.138
+
+### Patch Changes
+
+-   Updated dependencies [29179b5f]
+    -   @sap-ux/fiori-freestyle-writer@0.19.16
+    -   @sap-ux/fiori-elements-writer@0.12.24
+
 ## 0.2.137
 
 ### Patch Changes