+++ packages/ui5-application-writer/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @sap-ux/ui5-application-writer
 
+## 0.25.1
+
+### Patch Changes
+
+-   b63aea83: Update ui5 yaml spec to 3.1
+
+## 0.25.0
+
+### Minor Changes
+
+-   de0b13f2: Update templates to use Typescript version 5
+
+## 0.24.3
+
+### Patch Changes
+
+-   44df3d5c: fix lint warnings in locate-reuse-libs.js
+
+## 0.24.2
+
+### Patch Changes
+
+-   29179b5f: Add SAP icon loading in flpsandbox.html
+
 ## 0.24.1
 
 ### Patch Changes